Added DBHelper, NotebookDAO classes and test to insert

+++ Everpobre/app/src/main/java/com/ligartolabs/everpobre/model/Notebook.java
@@ -6,9 +6,9 @@ package com.ligartolabs.everpobre.model;
 public class Notebook {
 
     public static final String DEFAULT_NAME = "John Doe";
-    private int id;
+    private long id;
 
-    public Notebook(int id, String name) {
+    public Notebook(long id, String name) {
 
         if (name == null || name.isEmpty()) {
             this.name = DEFAULT_NAME;
@@ -23,11 +23,11 @@ public class Notebook {
         // disabled default constructor, customer required
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
